+++ src/main/clojure/com/github/clojure_repl/intellij/repl_command.clj
@@ -6,14 +6,19 @@
 
 (set! *warn-on-reflection* true)
 
-(def ^:private project-type->command
-  {:lein ["lein"]
-   :clojure ["clojure"]
-   :babashka ["bb"]
-   :shadow-cljs ["npx" "shadow-cljs"]
-   :boot ["boot"]
-   :nbb ["nbb"]
-   :gradle ["./gradlew"]})
+(def ^:private windows-os?
+  (.contains (System/getProperty "os.name") "Windows"))
+
+(defn ^:private project-type->command
+  [project-type]
+  ({:lein (if windows-os? ["lein.bat"] ["lein"])
+    :clojure ["clojure"]
+    :babashka ["bb"]
+    :shadow-cljs ["npx" "shadow-cljs"]
+    :boot ["boot"]
+    :nbb ["nbb"]
+    :gradle  (if windows-os? ["cmd" "/c" "gradlew.bat"] ["./gradlew"])}
+   project-type))
 
 (def ^:private middleware-versions
   ;; TODO make version configurable in intellij settings
@@ -36,9 +41,6 @@
   (println cmd-and-args)
   (apply shell/sh cmd-and-args))
 
-(def ^:private windows-os?
-  (.contains (System/getProperty "os.name") "Windows"))
-
 (defn ^:private normalize-command
   "Return CLASSPATH-CMD, but with the EXEC expanded to its full path (if found).
 
@@ -56,7 +58,7 @@
   replace the EXEC. If not, it tries the same with pwsh.exe."
   [[exec & args :as cmd]]
   (if (and windows-os?
-           (#{"clojure" "lein"} exec)
+           (#{"clojure" "lein.bat"} exec)
            (not (locate-executable exec)))
     (if-let [up (some #(when-let [ps (locate-executable %)]
                          (when (= 0 (:exit (apply shell (psh-cmd ps "Get-Command" exec))))
@@ -65,6 +67,12 @@
       (into up args)
       cmd)
     cmd))
+
+(defn ^:private remove-start-colon
+  [s]
+  (if (string/starts-with? s ":")
+    (subs s 1)
+    s))
 
 (defn ^:private project-type->parameters [project-type aliases jvm-args]
   (flatten
@@ -82,7 +90,10 @@
                   (map #(str "-J" (first %) "=" (second %)) jvm-args))
                 "-Sdeps"
                 "{:deps {nrepl/nrepl {:mvn/version \"%nrepl/nrepl%\"} cider/cider-nrepl {:mvn/version \"%cider/cider-nrepl%\"}} :aliases {:cider/nrepl {:main-opts [\"-m\" \"nrepl.cmdline\" \"--middleware\" \"[cider.nrepl/cider-middleware]\"]}}}"
-                (str "-M:" (string/join ":" (conj aliases "cider/nrepl")))]
+                (str "-M:" (->> "cider/nrepl"
+                                (conj (vec aliases))
+                                (map remove-start-colon)
+                                (string/join ":")))]
       :babashka ["nrepl-server" "localhost:0"]
       :shadow-cljs ["server"]
       :boot ["repl" "-s" "-b" "localhost" "wait"]